requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
google-generativeai
torch
transformers
Pillow
sentence-transformers
faiss-cpu
numpy